+++ web/src/actions/editApd/symbols.js
@@ -1,0 +1,5 @@
+export const ADD_APD_ITEM = Symbol('add apd item');
+export const ADD_APD_YEAR = Symbol('add apd year');
+export const EDIT_APD = Symbol('edit apd');
+export const REMOVE_APD_ITEM = Symbol('remove apd item');
+export const REMOVE_APD_YEAR = Symbol('remove apd year');
